bug #92460: tapebridged should gracefully shutdown a migration tape-session when tapegatewayd reports a disabled tape
Had to modify the unit-tests so they compile on SLC5 (I was previously using a
mac and I've just discovered it's file-system is NOT case-sensitive with respect
to filenames).

+++ test/unittest/castor/tape/tapebridge/BridgeProtocolEngineTest.hpp
@@ -898,11 +898,11 @@ public:
         CPPUNIT_ASSERT_EQUAL_MESSAGE(
           "Check maxFiles of request for more migration-work",
           bulkRequestConfigParams.getBulkRequestMigrationMaxFiles().getValue(),
-          moreWorkRequest->maxFiles());
+          (uint64_t)moreWorkRequest->maxFiles());
         CPPUNIT_ASSERT_EQUAL_MESSAGE(
           "Check maxBytes of request for more migration-work",
           bulkRequestConfigParams.getBulkRequestMigrationMaxBytes().getValue(),
-          moreWorkRequest->maxBytes());
+          (uint64_t)moreWorkRequest->maxBytes());
       }
 
       // Act as the client and send back the first file to be migrated
